add interactions binary for command management through terminal
