+++ discord-frontend/interactions/Cargo.toml
@@ -13,3 +13,15 @@ rust-version = "1.62.0"
 [dependencies]
 base = { path = "../../common/base", features = ["discord-model"] }
 ext = { path = "../ext", features = ["discord-model-ext"] }
+
+clap = { version = "3.2.15", features = ["cargo"] }
+tokio = { version = "1.20.1", features = ["macros", "rt", "rt-multi-thread"] }
+
+[[bin]]
+name = "interactions"
+path = "bin/main.rs"
+
+[lib]
+name = "interactions"
+crate-type = ["rlib"]
+path = "lib/lib.rs"
